// frontend/lib/api/generated/roundPublishRequestInitialStatus.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Asya Backend API
 * API documentation for Asya Backend application
 * OpenAPI spec version: @appVersion@
 */

/**
 * 初始状态
 */
export type RoundPublishRequestInitialStatus =
  (typeof RoundPublishRequestInitialStatus)[keyof typeof RoundPublishRequestInitialStatus];

export const RoundPublishRequestInitialStatus = {
  RUNNING: "RUNNING",
  PAUSED: "PAUSED",
} as const;